requirements.txt
flask
langchain
openai
faiss-cpu
sentence-transformers
transformers
PyPDF2
tiktoken
psycopg2-binary
sqlalchemy
python-dotenv
langchain
langchain-text-splitters
langchain-community
faiss-cpu
openai
sentence-transformers
transformers
flask
PyPDF2
python-dotenv

